fix: to pass all uni tests and delete segment fault.
Move storage map which is used in test suites to
OptimPredictorShutdownSpec.
