+++ spark/dl/src/main/scala/com/intel/analytics/bigdl/optim/ValidationMethod.scala
@@ -176,7 +176,8 @@ class Top1Accuracy[T: ClassTag](
     var count = 0
 
     val _target = target.asInstanceOf[Tensor[T]]
-    val _output = if (output.toTensor[T].size().head != _target.size().head) {
+    val _output = if (output.toTensor[T].nDimension() != 1 &&
+      output.toTensor[T].size().head != _target.size().head) {
       output.toTensor[T].narrow(1, 1, _target.size().head)
     } else {
       output.toTensor[T]
@@ -226,7 +227,8 @@ class Top5Accuracy[T: ClassTag](
   AccuracyResult = {
     var _target = target.asInstanceOf[Tensor[T]].squeezeNewTensor()
 
-    val _output = if (output.toTensor[T].size(1) != _target.size(1)) {
+    val _output = if (output.toTensor[T].nDimension() != 1 &&
+      output.toTensor[T].size(1) != _target.size(1)) {
       output.toTensor[T].narrow(1, 1, _target.size().head)
     } else {
       output.toTensor[T]
