+++ spark/dl/src/test/scala/com/intel/analytics/bigdl/optim/LocalOptimizerSpec.scala
@@ -456,35 +456,3 @@ class LocalOptimizerSpec extends FlatSpec with Matchers with BeforeAndAfter{
     assert(expectedG.almostEqual(newG, 0.0), "clipbynorm2 should generate correct gradient")
   }
 }
-
-@com.intel.analytics.bigdl.tags.Serial
-class LocalOptimizerSpec2 extends FlatSpec with Matchers with BeforeAndAfter {
-
-  import LocalOptimizerSpecModel._
-  import DummyDataSet._
-
-  before {
-    System.setProperty("bigdl.localMode", "true")
-    System.setProperty("bigdl.engineType", "mkldnn")
-    Engine.init
-  }
-
-  after {
-    System.clearProperty("bigdl.localMode")
-    System.clearProperty("bigdl.engineType")
-  }
-
-  "Train model and shutdown" should "be good" in {
-    RandomGenerator.RNG.setSeed(1000)
-    val model = dnnModel
-    val count = DnnStorage.get().count(!_._2)
-    val optimizer = new LocalOptimizer[Float](
-      model,
-      creDataSet,
-      new CrossEntropyCriterion[Float].asInstanceOf[Criterion[Float]]
-    ).setEndWhen(Trigger.severalIteration(1))
-
-    optimizer.optimize()
-    DnnStorage.get().count(!_._2) should be (count)
-  }
-}
